+++ CHANGES.md
@@ -1,5 +1,6 @@
 # vNext
 * Use `Zlib` from camlzip rather than cryptokit + crc
+* Allow capping the size of decompressed content
 
 # v0.1.0
 * Initial release